refactor breakpoint search into its own app
