+++ breakpoint_search/templates/breakpoint_search.html
@@ -71,7 +71,7 @@
 
     <script src="https://cdn.datatables.net/1.10.10/js/jquery.dataTables.min.js"></script>
     <link rel="stylesheet" href="https://cdn.datatables.net/1.10.10/css/jquery.dataTables.min.css"/>
-    <script type='text/javascript' src='{% static "js/family/breakpoint_search.js" %}'></script>
+    <script type='text/javascript' src='{% static "js/breakpoint_search.js" %}'></script>
 
     <iframe style='display:none;' name='igv' id='igv' src='about:blank'></iframe>
