+++ xbrowse_server/urls.py
@@ -14,6 +14,8 @@ import django.contrib.admindocs.urls
 import django.views.static
 import xbrowse_server.api.urls
 import seqr.urls
+import breakpoint_search
+
 
 admin.autodiscover()
 
@@ -108,10 +110,6 @@ urlpatterns = [
     url(r'^project/(?P<project_id>[\w.|-]+)/family/(?P<family_id>[\w.|-]+)/cause$', xbrowse_server.base.views.family_views.edit_family_cause, name='edit_family_cause'),
     url(r'^project/(?P<project_id>[\w.|-]+)/family/(?P<family_id>[\w.|-]+)/pedigreeimage/delete', xbrowse_server.base.views.family_views.pedigree_image_delete, name='pedigree_image_delete'),
 
-    # Breakpoint search
-    url(r'^project/(?P<project_id>[\w.|-]+)/family/(?P<family_id>[\w.|-]+)/breakpoint-search', xbrowse_server.base.views.family_views.breakpoint_search, name='breakpoint_search'),
-    url(r'^project/(?P<project_id>[\w.|-]+)/family/(?P<family_id>[\w.|-]+)/breakpoints', xbrowse_server.base.views.family_views.breakpoints, name='breakpoints'),
-    url(r'^project/(?P<project_id>[\w.|-]+)/breakpoint/(?P<breakpoint_id>[\w.|-]+)', xbrowse_server.base.views.project_views.project_breakpoint, name='project_breakpoint'),
 
     #
     # Cohort views
@@ -196,6 +194,10 @@ urlpatterns = [
     url(r'^matchmaker/search/project/(?P<project_id>[\w.|-]+)/family/(?P<family_id>[\w|-]+)$', xbrowse_server.matchmaker.views.matchmaker_search_page, name='matchmaker_search_page'), 
     url(r'^matchmaker/disclaimer$', xbrowse_server.matchmaker.views.matchmaker_disclaimer_page, name='matchmaker_disclaimer_page'), 
     url(r'^matchmaker/translate/matchbox_id$', xbrowse_server.matchmaker.views.matchbox_id_info, name='matchbox_id_info'),
+
+    # Breakpoint Search
+    url(r'', include('breakpoint_search.urls')),
+
 ]
 
 if settings.DEBUG != 4:
